refactor: User 도메인 리팩터링 외
- 소셜 로그인 서비스 주체는 SocialProvider enum으로 관리
- baseEntity는 global 바로 하위에 위치하도록
