+++ src/main/java/com/ll/commars/domain/auth/google/controller/ApiV1AuthGoogleController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ll.commars.domain.auth.auth.service.AuthService;
 import com.ll.commars.domain.auth.google.service.GoogleService;
-import com.ll.commars.domain.user.user.entity.User;
+import com.ll.commars.domain.user.entity.User;
 
 import lombok.RequiredArgsConstructor;
