+++ src/main/java/com/ll/commars/domain/auth/kakao/controller/ApiV1AuthKakaoController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ll.commars.domain.auth.auth.service.AuthService;
 import com.ll.commars.domain.auth.kakao.service.KakaoService;
-import com.ll.commars.domain.user.user.entity.User;
+import com.ll.commars.domain.user.entity.User;
 
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
