+++ src/main/java/com/ll/commars/domain/community/post/entity/Post.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.hibernate.annotations.ColumnDefault;
 
 import com.ll.commars.domain.community.comment.entity.Comment;
-import com.ll.commars.domain.user.user.entity.User;
-import com.ll.commars.global.baseEntity.BaseEntity;
+import com.ll.commars.domain.user.entity.User;
+import com.ll.commars.global.BaseEntity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
