+++ src/main/java/com/ll/commars/domain/favorite/favoriteRestaurant/entity/FavoriteRestaurant.java
@@ -1,9 +1,8 @@
-package com.ll.commars.domain.restaurant.menu.entity;
+package com.ll.commars.domain.favorite.favoriteRestaurant.entity;
 
 import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.global.BaseEntity;
+import com.ll.commars.domain.favorite.favorite.entity.Favorite;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -12,7 +11,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,32 +18,24 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "restaurants_menus")
+@Table(name = "favorites_restaurants")
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RestaurantMenu extends BaseEntity {
+public class FavoriteRestaurant {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	// 메뉴 이름
-	@NotNull
-	@Column(name = "name", nullable = false)
-	private String name;
+	// FavoriteRestaurant와 Favorite: 다대일
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "favorite_id")
+	private Favorite favorite;
 
-	@NotNull
-	@Column(name = "price", nullable = false)
-	private Integer price;
-
-	// 메뉴 사진
-	@Column(name = "image_url")
-	private String imageUrl;
-
-	// RestaurantMenu와 Restaurant: 다대일
+	// FavoriteRestaurant와 Restaurant: 다대일
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "restaurant_id")
 	private Restaurant restaurant;
