+++ src/main/java/com/ll/commars/domain/favorite/favoriteRestaurant/service/FavoriteRestaurantService.java
@@ -1,0 +1,24 @@
+package com.ll.commars.domain.favorite.favoriteRestaurant.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.ll.commars.domain.favorite.favoriteRestaurant.repository.FavoriteRestaurantRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class FavoriteRestaurantService {
+
+	private final FavoriteRestaurantRepository favoriteRestaurantRepository;
+
+	public void truncate() {
+		favoriteRestaurantRepository.deleteAll();
+	}
+
+	@Transactional
+	public void deleteRestaurantFromFavorite(long favoriteId, long restaurantId) {
+		favoriteRestaurantRepository.deleteByFavoriteIdAndRestaurantId(favoriteId, restaurantId);
+	}
+}
