+++ src/main/java/com/ll/commars/domain/restaurant/businessHour/entity/BusinessHour.java
@@ -3,7 +3,7 @@ package com.ll.commars.domain.restaurant.businessHour.entity;
 import java.time.LocalDateTime;
 
 import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.global.baseEntity.BaseEntity;
+import com.ll.commars.global.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
