+++ src/main/java/com/ll/commars/domain/review/review/entity/Review.java
@@ -1,8 +1,8 @@
 package com.ll.commars.domain.review.review.entity;
 
 import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.domain.user.user.entity.User;
-import com.ll.commars.global.baseEntity.BaseEntity;
+import com.ll.commars.domain.user.entity.User;
+import com.ll.commars.global.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
