+++ src/main/java/com/ll/commars/domain/user/entity/Gender.java
@@ -1,0 +1,8 @@
+package com.ll.commars.domain.user.entity;
+
+public enum Gender {
+
+	남자,
+	여자,
+	;
+}
