+++ src/main/java/com/ll/commars/domain/user/entity/SocialProvider.java
@@ -1,0 +1,9 @@
+package com.ll.commars.domain.user.entity;
+
+public enum SocialProvider {
+
+	GOOGLE,
+	KAKAO,
+	NAVER,
+	;
+}
