--- src/main/java/com/ll/commars/domain/user/favoriteRestaurant/entity/FavoriteRestaurant.java
@@ -1,42 +1,0 @@
-package com.ll.commars.domain.user.favoriteRestaurant.entity;
-
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.domain.user.favorite.entity.Favorite;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "favorites_restaurants")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class FavoriteRestaurant {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	// FavoriteRestaurant와 Favorite: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "favorite_id")
-	private Favorite favorite;
-
-	// FavoriteRestaurant와 Restaurant: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "restaurant_id")
-	private Restaurant restaurant;
-}
