+++ src/main/java/com/ll/commars/domain/user/repository/UserRepository.java
@@ -1,0 +1,17 @@
+package com.ll.commars.domain.user.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ll.commars.domain.user.entity.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+	Optional<User> findByEmail(String email);
+
+	Optional<User> findByEmailAndName(String email, String name);
+
+	Optional<User> findByIdAndEmail(Long id, String email);
+
+}
