+++ src/main/java/com/ll/commars/global/jwt/component/JwtProvider.java
@@ -9,7 +9,7 @@ import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.ll.commars.domain.user.user.entity.User;
+import com.ll.commars.domain.user.entity.User;
 import com.ll.commars.global.jwt.entity.JwtToken;
 
 import io.jsonwebtoken.Claims;
